+++ deode/suites/discover_suite.py
@@ -1,5 +1,4 @@
-"""Discover tasks."""
-import contextlib
+"""Discover suites."""
 import importlib
 import inspect
 import os
@@ -7,9 +6,9 @@ import pkgutil
 import sys
 import types
 
-from ..logs import LoggerHandlers, logger
+from ..logs import logger
 from ..plugin import DeodePluginRegistryFromConfig
-from .base import Task, _get_name
+from .base import SuiteDefinition, _get_name
 
 
 def discover_modules(package, what="plugin"):
@@ -27,12 +26,12 @@ def discover_modules(package, what="plugin"):
 
     """
     path = package.__path__
-    prefix = package.__name__ + ".tasks."
+    prefix = package.__name__ + ".suites."
 
-    logger.info("{} search path: {}", what.capitalize(), path)
+    logger.debug("{} search path: {}", what.capitalize(), path)
     for _finder, mname, _ispkg in pkgutil.iter_modules(path):
         fullname = prefix + mname
-        logger.info("Loading module {}", fullname)
+        logger.debug("Loading module {}", fullname)
         try:
             mod = importlib.import_module(fullname)
         except ImportError as exc:
@@ -41,64 +40,60 @@ def discover_modules(package, what="plugin"):
         yield fullname, mod
 
 
-def get_task(name, config):
-    """Create a `deode.tasks.Task` object from configuration.
+def get_suite(name, config):
+    """Create a `deode.suites.SuiteDefinition` object from configuration.
 
     Args:
         name (_type_): _description_
         config (_type_): _description_
 
     Raises:
-        NotImplementedError: If task `name` is not amongst the known task names.
+        NotImplementedError: If SuiteDefinition `name` is not amongst
+                             the known SuiteDefinition names.
 
     Returns:
         _type_: _description_
     """
-    with contextlib.suppress(KeyError):
-        # loglevel may have been overridden, e.g., via ECFLOW UI
-        logger.configure(
-            handlers=LoggerHandlers(default_level=config["general.loglevel"])
-        )
-        logger.debug("Logger reset to level {}", config["general.loglevel"])
-
     reg = DeodePluginRegistryFromConfig(config)
-    known_types = available_tasks(reg)
+    known_types = available_suites(reg)
     try:
         cls = known_types[name.lower()]
     except KeyError as error:
-        raise NotImplementedError(f'Task "{name}" not implemented') from error
+        raise NotImplementedError(f'SuiteDefinition "{name}" not implemented.') from error
 
     return cls(config)
 
 
-def available_tasks(reg):
+def available_suites(reg):
     """Create a list of available tasks.
 
     Args:
         reg (DeodePluginRegistry): Deode plugin registry
 
     Returns:
-        known_types (list): Task objects
+        known_types (list): Suite objects
 
     """
     known_types = {}
     for plg in reg.plugins:
-        if os.path.exists(plg.tasks_path):
-            tasks = types.ModuleType(plg.name)
-            tasks.__path__ = [plg.tasks_path]
+        logger.info("Check {} with path {}", plg.name, plg.suites_path)
+        if os.path.exists(plg.suites_path):
+            suites = types.ModuleType(plg.name)
+            suites.__path__ = [plg.suites_path]
+            logger.info("Adding plugin {} with path {}", plg.name, plg.suites_path)
             sys.path.insert(0, plg.path)
-            found_types = discover(tasks, Task)
+            found_types = discover(suites, SuiteDefinition)
             for ftype, cls in found_types.items():
                 if ftype in known_types:
                     logger.warning("Overriding suite {}", ftype)
                 known_types[ftype] = cls
         else:
-            logger.warning("Plug-in task {} not found", plg.tasks_path)
+            logger.warning("Plug-in task {} not found", plg.suites_path)
     return known_types
 
 
 def discover(package, base):
-    """Discover task classes.
+    """Discover SuiteDefinition classes.
 
     Plugin classes are discovered in a given namespace package, deriving from
     a given base class. The base class itself is ignored, as are classes
@@ -114,24 +109,21 @@ def discover(package, base):
     Returns:
         (dict of str: type): Discovered plugin classes
     """
-    what = base.__name__
 
     def pred(x):
         return inspect.isclass(x) and issubclass(x, base) and x is not base
 
     discovered = {}
-    for fullname, mod in discover_modules(package, what=what):
+    for fullname, mod in discover_modules(package):
+        logger.debug("fullname={} mod={}", fullname, mod)
         for cname, cls in inspect.getmembers(mod, pred):
-            tname = _get_name(cname, cls, what.lower())
+            tname = _get_name(cname, cls)
+            logger.debug("tname={}", tname, cname)
             if cls.__module__ != fullname:
-                logger.info(
-                    "Skipping {} %r imported by %r", what.lower(), tname, fullname
-                )
+                logger.debug("Skipping {} imported by {}", tname, fullname)
                 continue
             if tname in discovered:
-                logger.warning(
-                    "{} type %r is defined more than once", what.capitalize(), tname
-                )
+                logger.warning("type {} is defined more than once", tname)
                 continue
             discovered[tname] = cls
     return discovered